requirements.txt
psycopg2-binary>=2.9.9
openai>=1.12.0
moviepy>=2.0.0
matplotlib>=3.8.0
numpy>=1.26.0
Pillow>=10.2.0
python-dotenv>=1.0.1
gTTS>=2.5.1
requests>=2.31.0
